Update README.md
Suppose a teacher has chocolate bars in the sizes such that the minwidth and maxwidth and of minheight and maxheight. The teacher has to distribute the chocolates to the students only when the chocolate is a square.that means suppose if she has the chocolate bars of minwidth=2 and maxwidth=4 and minheight=3 and maxheight=6, then she has bars of sizes (2,3),(2,4),(2,5),(2,6),(3,3),(3,4),(3,5),(3,6),(4,3),(4,4),(4,5),(4,6);  she can give the chocolate to a student if the width and height of the chocolate is same, and if not then she will break the chocolate to make it square such as (2,3) can be broken down into (2,2) and(2,1) and later (2,1)will be broken down into (1,1) and (1,1).

eg.
input
1
3
1
2
output 
12
how many kids will get the chocolate for given input?